requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu111
--trusted-host download.pytorch.org
alembic==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
antlr4-python3-runtime==4.8
anyio==3.6.1; python_full_version >= '3.6.2'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0; python_version >= '3.6'
asttokens==2.0.5
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
babel==2.10.3; python_version >= '3.6'
backcall==0.2.0
beautifulsoup4==4.11.1; python_version >= '3.6'
bleach==5.0.1; python_version >= '3.7'
captum==0.4.1
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
cloudpickle==2.1.0; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
databricks-cli==0.17.0
debugpy==1.6.0; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
docker==5.0.3; python_version >= '3.6'
entrypoints==0.4; python_version >= '3.6'
executing==0.8.3
fastjsonschema==2.15.3
flask==2.1.2; python_version >= '3.7'
fonttools==4.33.3; python_version >= '3.7'
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.27; python_version >= '3.7'
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==20.1.0; platform_system != 'Windows'
hydra-core==1.1.1
idna==3.3; python_version >= '3.5'
importlib-metadata==4.12.0; python_version < '3.10' and python_version < '3.10'
ipykernel==6.15.0; python_version >= '3.7'
ipython-genutils==0.2.0
ipython==8.4.0; python_version >= '3.8'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.18.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.1.0; python_version >= '3.6'
json5==0.9.8
jsonschema==4.6.1; python_version >= '3.7'
jupyter-client==7.3.4; python_version >= '3.7'
jupyter-core==4.10.0; python_version >= '3.7'
jupyter-server==1.18.0; python_version >= '3.7'
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-server==2.14.0; python_version >= '3.7'
jupyterlab==3.2.9
kiwisolver==1.4.3; python_version >= '3.7'
mako==1.2.1; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.3; python_version >= '3.5'
matplotlib==3.5.2; python_version >= '3.7'
mistune==0.8.4
mlflow==1.20.2
nbclassic==0.4.0; python_version >= '3.7'
nbclient==0.6.6; python_version >= '3.7'
nbconvert==6.5.0; python_version >= '3.7'
nbformat==5.4.0; python_version >= '3.7'
nest-asyncio==1.5.5; python_version >= '3.5'
notebook-shim==0.1.0; python_version >= '3.7'
numpy==1.23.0; python_version >= '3.9' and python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
oauthlib==3.2.0; python_version >= '3.6'
omegaconf==2.1.2; python_version >= '3.6'
opencv-python==4.5.5.62
packaging==21.3; python_version >= '3.6'
pandas==1.4.3; python_version >= '3.8'
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.2.0; python_version >= '3.7'
prometheus-client==0.14.1; python_version >= '3.6'
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.30; python_full_version >= '3.6.2'
protobuf==3.19.4
psutil==5.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0; os_name != 'nt'
pure-eval==0.2.2
pycparser==2.21
pygments==2.12.0; python_version >= '3.6'
pyjwt==2.4.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-editor==1.0.4
pytz==2022.1
pyyaml==6.0; python_version >= '3.6'
pyzmq==23.2.0; python_version >= '3.6'
querystring-parser==1.2.4
requests==2.28.1; python_version >= '3.7' and python_version < '4'
scikit-learn==1.0.2
scipy==1.8.1; python_version < '3.11' and python_version >= '3.8'
send2trash==1.8.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.0; python_version >= '3.6'
sniffio==1.2.0; python_version >= '3.5'
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlalchemy==1.4.39; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlparse==0.4.2; python_version >= '3.5'
stack-data==0.3.0
tabulate==0.8.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
terminado==0.15.0; python_version >= '3.7'
threadpoolctl==3.1.0; python_version >= '3.6'
tinycss2==1.1.1; python_version >= '3.6'
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tornado==6.1; python_version >= '3.5'
traitlets==5.3.0; python_version >= '3.7'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3; python_version >= '3.7'
werkzeug==2.1.2; python_version >= '3.7'
zipp==3.8.0; python_version >= '3.7'
